Adding Range Calculation into the Attack System
adding an if to check the distance of the current bot and the selected
bot when on attacking mode. The max range is calculated by using pre
defined number (for tank and troop) in inspector and times them with the
constant multiplier gridScale. gridScale need some fixing everytime the
scale of the game is changed (by ImageTarget scale difference)