Fixed bug where player would lose when they lost one bot.
Player Controllers CheckChildren method has a second if statement that checks if the "hasChildren" bool is true. This if statement was in a for each loop that was determining whether this should be set as true or false once the for each was complete. This if statement has now been removed from the for each.